Refactor input file eto
